+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ## Changelog
 ### V1.1.1
 * Extended `NumericSensor` and `NumericSensorActor` capabilities to support `Rollershutter` and `Dimmer` types.
+* Allowing user defined `minTemp`, `maxTemp` and `minStep` for `Thermostat` and `TemperatureSensor` accessory types.
 
 ### V1.1.0
 * Updated Readme to fix typo's